Development

+++ src/controllers/teams.controller.js
@@ -1,10 +1,15 @@
 const db = require("../models/index");
 const Teams = db.teams;
 
+// Done
 exports.findAll = (req, res) => {
   Teams.find()
     .then((result) => {
-      res.send(result);
+      res.send({
+        message: "Teams successfully fetched.",
+        timestamp: new Date().toString(),
+        data: result,
+      });
     })
     .catch((err) => {
       res.status(500).send({
@@ -13,12 +18,29 @@ exports.findAll = (req, res) => {
     });
 };
 
+// Done
 exports.findOne = (req, res) => {
   const id = req.params.id;
 
+  if (!id) {
+    return res.status(400).send({
+      message: "Team ID is required.",
+    });
+  }
+
   Teams.findById(id)
     .then((result) => {
-      res.send(result);
+      if (!result) {
+        res.status(404).send({
+          message: "Team not found",
+        });
+      }
+
+      res.send({
+        message: "Team successfully retrieved.",
+        timestamp: new Date().toString(),
+        data: result,
+      });
     })
     .catch((err) => {
       res.status(500).send({
@@ -27,8 +49,15 @@ exports.findOne = (req, res) => {
     });
 };
 
+// Done
 exports.delete = (req, res) => {
   const id = req.params.id;
+
+  if (!id) {
+    return res.status(400).send({
+      message: "Team ID is required.",
+    });
+  }
 
   Teams.findByIdAndRemove(id)
     .then((result) => {
@@ -39,7 +68,8 @@ exports.delete = (req, res) => {
       }
 
       res.send({
-        message: "Team was deleted",
+        message: "Team successfully deleted.",
+        timestamp: new Date().toString(),
       });
     })
     .catch((err) => {
@@ -49,14 +79,23 @@ exports.delete = (req, res) => {
     });
 };
 
+// Done
 exports.create = (req, res) => {
+  const { name, description, position, email } = req.body;
+
+  if (!name || !description || !position || !email) {
+    return res.status(400).send({
+      message: "Please fill all required fields.",
+    });
+  }
+
   const team = new Teams({
-    name: req.body.name,
-    position: req.body.position,
-    photo: req.file.path,
+    name: name,
+    description: description,
+    position: position,
     contact: {
       instagram: req.body.instagram,
-      facebook: req.body.facebook,
+      email: email,
       twitter: req.body.twitter,
       linkedin: req.body.linkedin,
     },
@@ -67,11 +106,89 @@ exports.create = (req, res) => {
     .then((result) => {
       res.status(200).send({
         message: "Team successfully added.",
+        timestamp: new Date().toString(),
+        data: result,
       });
     })
     .catch((err) => {
-      res.status(409).send({
+      res.status(500).send({
         message: err.message || "Some error while creating team.",
+        timestamp: new Date().toString(),
+      });
+    });
+};
+
+// Done
+exports.update = (req, res) => {
+  const id = req.params.id;
+
+  if (!id) {
+    res.status(400).send({
+      message: "Teams ID is required",
+    });
+  }
+
+  Teams.findByIdAndUpdate(id, req.body)
+    .then((result) => {
+      if (!result) {
+        res.status(404).send({
+          message: "Team not found",
+        });
+      }
+
+      res.send({
+        message: "Team was updated",
+        timestamp: new Date().toString(),
+      });
+    })
+    .catch((err) => {
+      res.status(500).send({
+        message: err.message || "Some error while updating team.",
+        timestamp: new Date().toString(),
+      });
+    });
+};
+
+// Done
+exports.teamProfile = (req, res) => {
+  const id = req.params.id;
+
+  if (!id) {
+    res.status(400).send({
+      message: "Teams ID is required",
+    });
+  }
+
+  if (!req.file) {
+    return res.status(422).send({
+      message: "Team profile photo is required.",
+    });
+  }
+
+  const photoName = req.file.filename;
+  const photoLink = `${req.protocol}://${req.get(
+    "host"
+  )}/assets/foto/${photoName}`;
+
+  Teams.findByIdAndUpdate(id, { photo: { photoName, photoLink } })
+    .then((result) => {
+      if (!result) {
+        res.status(404).send({
+          message: "Team not found",
+        });
+      }
+
+      res.send({
+        message: "Team profile photo successfully updated.",
+        timestamp: new Date().toString(),
+      });
+    })
+    .catch((err) => {
+      res.status(500).send({
+        message:
+          err.message ||
+          "Some error occurred while changing the profile picture.",
+        timestamp: new Date().toString(),
       });
     });
 };
